Enemy attack cooldown and resume attack almost works.
Somehow, the lose health function is being called from the player's body entered signal when the enemy's cooldown times out, and I don't know why. It should get called once and only once during the time that the enemy's mouth shape is overlapping the player, so I'm not exactly sure what's going on.